+++ CHANGELOG.md
@@ -7,6 +7,7 @@ We follow [Calendar Versions](https://calver.org/).
 - Add PC Analysis
 - Add time series tools
 - Add apply function for lowpass and bandstop time-filter to field
+- Add function to open tracmass files
 
 ## Version 2021.2